pyproject.toml
[tool.poetry]
name = "quant-api"
version = "0.1.0"
description = ""
authors = ["Hongmin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
httpx = "0.27.2"
pydantic = "^2.10.2"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
pyyaml = "^6.0.2"
numpy = "^2.1.3"
pandas = "^2.2.3"
aiosqlite = "^0.20.0"
websockets = "^14.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"
pytest-html = "^4.1.1"
greenlet = "^3.1.1"
httpx-ws = "^0.6.2"
python-multipart = "^0.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
